+++ src/main/java/teamssavice/ssavice/user/service/UserReadService.java
@@ -1,0 +1,25 @@
+package teamssavice.ssavice.user.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import teamssavice.ssavice.global.constants.ErrorCode;
+import teamssavice.ssavice.global.exception.EntityNotFoundException;
+import teamssavice.ssavice.user.entity.Users;
+import teamssavice.ssavice.user.infrastructure.repository.UserRepository;
+
+import java.util.Optional;
+
+@Service
+@RequiredArgsConstructor
+public class UserReadService {
+    private final UserRepository userRepository;
+
+    public Optional<Users> findByEmail(String email) {
+        return userRepository.findByEmail(email);
+    }
+
+    public Users findById(Long id) {
+        return userRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(ErrorCode.USER_NOT_FOUND));
+    }
+}
